% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{Annotate}
\alias{Annotate}
\title{Gene annotation}
\usage{
Annotate(dataset, id.type = ID.TYPE.geneSymbol, platform = "",
  species = "", annotation.file = NA)
}
\arguments{
\item{dataset}{A numeric matrix of gene expressions or a Study
object. For matrix, Each row is for one gene and each column is
for one sample. The rownames should be gene IDs, which could
be probe IDs, refseq IDs, entrez IDs, or gene symbols.
For Study object, the same rule for matrix apply to each matrix
in the Study@datasets list.}

\item{id.type}{a character string to specify the type of gene
ID. It should be \code{"ProbeID"}, \code{"GeneSymbol"},
\code{"RefSeqID"}, or \code{"EntrezID"}.}

\item{platform}{a character string to specify the platform of the
gene expression data set. See details.}

\item{species}{a character string to specify the species of the
gene expression data set. The accepted \code{species} include
\code{"human"}, \code{"mouse"}, and \code{"rat"}.}

\item{annotation.file}{a character string of a txt file name to
upload user-specific annotation file. The default is
\code{NA}. See details.}
}
\value{
If dataset is matrix, it returns A numeric matrix of gene
expressions. with row names being genesymbol. If dataset is Study
object, a Study object is returned with Study@datasets annotates.
}
\description{
Annotation for the gene expression data set.
}
\details{
Annotate will transform the row names of the dataset to gene
symbol. The original row names could be either gene symbol
(in the case it is already gene symbol, no tranformation take
place), reference sequence, or entrez id.

For human, the accepted platforms include \code{"chip hgu95a"},
\code{"chip hgu95av2"}, \code{"chip hgu95b"}, \code{"chip
hgu95c"}, \code{"chip hgu95d"}, \code{"chip hgu95e"}, \code{"chip
hgu133a"}, \code{"chip hgu133b"}, \code{"hgu133plus2"},
\code{"Illumina HumanHT-12 V3 Bead Array"}, \code{"Illumina
HumanHT-12 V4 Bead Array"}, \code{"Illumina HumanWG-6 V1 Bead
Array"}, \code{"Illumina HumanWG-6 V2 Bead Array"},
\code{"Illumina HumanWG V2 Bead Array"}, \code{"chip hgug4100a"},
\code{"chip hgug4101a"}, \code{"chip hgug4110b"}, \code{"chip
hgug4111a"}, and \code{"chip hgug4112a"}.

For mouse, the accepted platforms include \code{"chip mgu74a"},
\code{"chip mgu74av2"}, \code{"chip mgu74b"}, \code{"chip
mgu74bv2"}, \code{"chip mgu74c"}, \code{"chip mgu74cv2"},
\code{"chip moe430a"}, \code{"chip moe430b"}, \code{"Affymetrix
GeneChip Mouse Genome 430 2.0"}, \code{"Illumina MouseWG-6 v1.0
Bead Array"}, \code{"Illumina MouseWG-6 v2.0 Bead Array"},
\code{"chip mgug4104a"}, \code{"chip mgug4120a"}, \code{"chip
mgug4121a"}, and \code{"chip mgug4122a"}.

For rat, the accepted platforms include \code{"chip rat2302"},
\code{"chip rgu34a"}, \code{"chip rgu34b"}, \code{"chip rgu34c"},
\code{"Illumina Rat v1.0 Bead Array"}, \code{"chip rgug4105a"},
\code{"chip rgug4130a"}, and \code{"chip rgug4131a"}.

For other platforms, an annotation file is needed to be uploaded
by \code{annotation.file}. The annotation file to be read should
be prepared strictly according following format. The first column
should be probe ID, and the second column should be gene
symbol. The first row should be the column headers.
}
\examples{
data(datasets.eg)
data(preproc.option)
# Annotate Expression Matrix
res <- Annotate(dataset=datasets.eg[[1]], id.type=ID.TYPE.probeID, 
                platform=PLATFORM.hgu133plus2)
# Annotate Study
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=datasets.eg[1])
res <- Annotate(study, id.type=ID.TYPE.probeID, platform=PLATFORM.hgu133plus2)
}
\author{
Lin Wang, Schwannden Kuo
}

