% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations_table.R
\name{CorrelationsTable}
\alias{CorrelationsTable}
\title{Generate one-vs-the-rest correlation metrics for a target variable}
\usage{
CorrelationsTable(data, target, method = "pearson", remove.na.cols = FALSE,
  verbose = TRUE)
}
\arguments{
\item{data}{A dataframe.}

\item{target}{str. The name of the column to correlate against.}

\item{method}{str. a character string indicating which correlation coefficient
(or covariance) is to be computed. One of "pearson" (default),
"kendall", or "spearman".}

\item{remove.na.cols}{if TRUE, remove columns containing NAs.}

\item{verbose}{if TRUE print names of removed columns.}
}
\value{
The original dataframe the specified columns removed.
}
\description{
Generate one-vs-the-rest correlation metrics for a target variable
}
