% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_columns.R
\name{DropColumns}
\alias{DropColumns}
\title{Drop columns from a dataframe}
\usage{
DropColumns(data, columns)
}
\arguments{
\item{data}{A dataframe.}

\item{columns}{vec. The names of the columns to drop}
}
\value{
The original dataframe the specified columns removed.
}
\description{
Drop columns from a dataframe
}
