% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{DTYPE.RNAseq.FPKM}
\alias{DTYPE.RNAseq.FPKM}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
DTYPE.RNAseq.FPKM
}
\description{
Constant Option
}
\keyword{datasets}

