% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{Filter}
\alias{Filter}
\title{Filter genes}
\usage{
Filter(datasets, data.type, del.perc = c(0.3, 0.3), threshold = 1)
}
\arguments{
\item{datasets}{a list of gene expression matrice. Each matrix is
for one study. Each row of the matrix is for one gene and each
column is for one sample. The row names are gene symbols.}

\item{data.type}{a character string to specify the type of data in
\code{datasets}. It should be \code{"microarray"},
\code{"RNAseq-FPKM"}, or \code{"RNAseq-count"}.}

\item{del.perc}{a numeric vector with two elements, which specify
the percentage of genes to be filtered in the two sequential steps
of gene filtering when \code{data.type} is \code{"microarray"} or
\code{"RNAseq-FPKM"}. The default is \code{c(0.3, 0.3)}. See Details.}

\item{threshold}{a numeric value to specify the threshold when
\code{data.type} is \code{"RNAseq-count"}. The default is \code{1}.
See details.}
}
\value{
A list of gene expression matrice after filtering. Each
matrix is for one study. Each row of the matrix is for one gene
and each column is for one sample. The row names are gene
symbols.
}
\description{
Filter genes with low means and low variances.
}
\details{
When \code{data.type} is \code{"microarray"} or
\code{"RNAseq-FPKM"}, two sequential steps of gene filtering are
performed. In the first step, the genes with very low expressions
are filtered out. These genes are identified with small average
expression values across studies. Specifically, mean intensities
of each gene across all samples in each study are calculated and
the corresponding ranks are obtained. The sum of such ranks across
studies of each gene is calculated and genes with the lowest
\code{del.perc[1]} percent rank sum are considered un-expressed
genes (i.e. small expression intensities) and filtered
out. Similarly, in the second step, the non-informative (small
variation) genes are filtered out by replacing mean intensity in
the first step with standard deviation. Genes with the lowest
\code{del.perc[2]} percent rank sum of standard deviations are
filtered out.

When \code{data.type} is \code{"RNAseq-count"}, the genes with
very low counts are filtered out. These genes are identified with
minimum of mean counts across studies.
}
\examples{
data(study.eg)
data(preproc.option)
SinglePreproc <- function(x) {
  x <- Annotate(dataset=x, id.type = "ProbeID", platform=PLATFORM.hgu133plus2)
  x <- Impute(dataset=x)
  x <- PoolReplicate(dataset=x)
}
study.eg <- lapply(study.eg, SinglePreproc)
study.eg <- Merge(datasets=study.eg)
res <- Filter(datasets=study.eg, data.type=DTYPE.microarray, del.perc=c(0.3, 0.2))
}
\author{
Lin Wang
}

