% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{ID.TYPE.EntrezID}
\alias{ID.TYPE.EntrezID}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
ID.TYPE.EntrezID
}
\description{
Constant Option
}
\keyword{datasets}

