% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\docType{methods}
\name{Impute,matrix-method}
\alias{Impute,matrix-method}
\title{Title Impute Gene Expression Matrix}
\usage{
\S4method{Impute}{matrix}(dataset, method = IMPUTE.method.knn, k = 10)
}
\arguments{
\item{dataset}{matrix. A gene expression matrix.}

\item{method}{a character string to specify how to deal with the
missing values. It could be \code{"impute"} or \code{"remove"}, which
means to impute missing values using nearest
neighbor averaging or remove the genes with missing values
directly.}

\item{k}{the number of neighbors to be used in the imputation. The
default is \code{10}.}
}
\value{
A gene expression matrix with missing value imputed.
}
\description{
Title Impute Gene Expression Matrix
}
\examples{
data(study.eg)
data(preproc.option)
res <- Impute(study.eg[[1]])
}

