% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\docType{methods}
\name{Impute,Study-method}
\alias{Impute,Study-method}
\title{Title Impute Study}
\usage{
\S4method{Impute}{Study}(dataset, method = IMPUTE.method.knn, k = 10)
}
\arguments{
\item{dataset}{Study. A Study object.}

\item{method}{a character string to specify how to deal with the
missing values. It could be \code{"impute"} or \code{"remove"}, which
means to impute missing values using nearest
neighbor averaging or remove the genes with missing values
directly.}

\item{k}{the number of neighbors to be used in the imputation. The
default is \code{10}.}
}
\value{
Study object with missing value imputed.
}
\description{
Title Impute Study
}
\author{
Schwannden Kuo
}

