% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{IMPUTE.method.knn}
\alias{IMPUTE.method.knn}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
IMPUTE.method.knn
}
\description{
Constant Option
}
\keyword{datasets}

