% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{Impute}
\alias{Impute}
\title{Impute missing values}
\usage{
Impute(dataset, method = IMPUTE.method.knn, k = 10)
}
\arguments{
\item{dataset}{a numeric matrix of gene expressions or a Study object.
For matrix, each row is for one gene and each column is for one sample.
The row names are gene symbols. For Study object, the same rule applies
to matrix in Study@datasets}

\item{method}{a character string to specify how to deal with the
missing values. It could be \code{"impute"} or \code{"remove"}, which
means to impute missing values using nearest
neighbor averaging or remove the genes with missing values
directly.}

\item{k}{the number of neighbors to be used in the imputation. The
default is \code{10}.}
}
\value{
A matrix of the gene expressions after dealing with
missing values.
}
\description{
Impute missing values in gene expression data set using nearest neighbor
averaging or remove the genes with missing values directly.
}
\details{
See also \link{knnImputation}.
}
\examples{
data(datasets.eg)
data(preproc.option)
# Impute for expression matrix
res <- Impute(datasets.eg[[1]], method=IMPUTE.method.knn)
# Impute for Study object
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=datasets.eg[1])
res <- Impute(study, method=IMPUTE.method.remove)
}
\author{
Lin Wang, Schwannden Kuo
}

