% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{is.continuous}
\alias{is.continuous}
\title{Title is.discrete}
\usage{
is.continuous(studies)
}
\arguments{
\item{studies}{studies vector of string (data type) or list of studies.}
}
\value{
a vector with each entry corresponding to whether the data type or 
study is continuous.
}
\description{
Title is.discrete
}
\examples{
data(preproc.option)
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=list(matrix()))
is.continuous(study)
}
\author{
Schwannden Kuo
}

