% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{Merge}
\alias{Merge}
\title{Merge microarray data sets}
\usage{
Merge(datasets, name)
}
\arguments{
\item{datasets}{a list of gene expression matrice or list of Study. 
object. Each row of the matrix (study@datasets[[i]]) is for one gene,
each column is for one sample. The row names are gene symbols.}

\item{name}{The name od the Study object that will be returened}
}
\value{
If dataset is a list of gene expression matrice, it returns 
a list of gene expression matrice after merging. Each matrix is for 
one study. Each row of the matrix is for one gene and each column 
is for one sample. The row names are gene symbols.
If dataset is a list of Study object, it returns a Study object with
multiple study type, and datasets would be a list of gene expression 
matrice after merging, clinicals would be a list of clinical data 
after merging.
}
\description{
Merge microarray data sets in possibly irregular order.
}
\details{
The gene expression data sets may be in possibly irregular order
with different numbers of genes. This function is used to extract
the common genes across studies. The merged data sets have the
same genes in the same order.
}
\examples{
data(study.eg)
data(preproc.option)
SinglePreproc <- function(x) {
  x <- Annotate(dataset=x, id.type=ID.TYPE.probeID, platform=PLATFORM.hgu133plus2)
  x <- Impute(dataset=x)
  x <- PoolReplicate(dataset=x)
}
study.eg <- lapply(study.eg, SinglePreproc)
res <- Merge(datasets = study.eg)
}
\author{
Lin Wang, Schwannden Kuo
}

