% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{meta}
\alias{meta}
\title{Title Generic Function: meta}
\usage{
meta(object)
}
\arguments{
\item{object}{an object.}
}
\value{
The meta data of the object.
}
\description{
Title Generic Function: meta
}
\examples{
data(preproc.option)
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=list(matrix()))
meta(study)
}
\author{
Schwannden Kuo
}

