% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\docType{methods}
\name{ntype,character-method}
\alias{ntype,Study-method}
\alias{ntype,character-method}
\title{ntype method}
\usage{
\S4method{ntype}{character}(object)

\S4method{ntype}{Study}(object)
}
\arguments{
\item{object}{a Study object or a string in DTYPE.all.}
}
\description{
ntype method

ntype method
}

