% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{NTYPE.continuous}
\alias{NTYPE.continuous}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
NTYPE.continuous
}
\description{
Constant Option
}
\keyword{datasets}

