% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{ntype}
\alias{ntype}
\title{Title Generic Function: ntype}
\usage{
ntype(object)
}
\arguments{
\item{object}{a Study object or a string in DTYPE.all.}
}
\value{
the numeric type of the object.
}
\description{
Title Generic Function: ntype
}
\examples{
# ntype for Study
data(preproc.option)
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=list(matrix()))
ntype(study)
# ntype for dtype character
ntype(DTYPE.RNAseq.count)
}
\author{
Schwannden Kuo
}

