% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{PLATFORM.all}
\alias{PLATFORM.all}
\title{Constant Option}
\format{An object of class \code{character} of length 42.}
\usage{
PLATFORM.all
}
\description{
Constant Option
}
\keyword{datasets}

