% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{PLATFORM.GenePLATFORM.Mouse.Genome.430.2.0}
\alias{PLATFORM.GenePLATFORM.Mouse.Genome.430.2.0}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
PLATFORM.GenePLATFORM.Mouse.Genome.430.2.0
}
\description{
Constant Option
}
\keyword{datasets}

