% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{PLATFORM.mgu74a}
\alias{PLATFORM.mgu74a}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
PLATFORM.mgu74a
}
\description{
Constant Option
}
\keyword{datasets}

