% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{PLATFORM.MouseWG.6.v2.0.Bead.Array}
\alias{PLATFORM.MouseWG.6.v2.0.Bead.Array}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
PLATFORM.MouseWG.6.v2.0.Bead.Array
}
\description{
Constant Option
}
\keyword{datasets}

