% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.replicate.R
\docType{methods}
\name{PoolReplicate,matrix-method}
\alias{PoolReplicate,matrix-method}
\title{Title Handle Duplicated Gene Symbol}
\usage{
\S4method{PoolReplicate}{matrix}(dataset, method = REPLICATE.IQR)
}
\arguments{
\item{dataset}{matrix. A gene expression matrix.}

\item{method}{A character string of the method to deal with the
replicated gene symbols. It can be \code{"IQR"}, \code{"largest"}, or
\code{"average"}. The default is \code{"IQR"}. See details.}
}
\value{
A gene expression matrix with unique row entries.
}
\description{
Title Handle Duplicated Gene Symbol
}
\author{
Schwannden Kuo
}

