% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.replicate.R
\docType{methods}
\name{PoolReplicate,Study-method}
\alias{PoolReplicate,Study-method}
\title{Title Handle Duplicated Gene Symbol}
\usage{
\S4method{PoolReplicate}{Study}(dataset, method = REPLICATE.IQR)
}
\arguments{
\item{dataset}{Study. A Study object.}

\item{method}{A character string of the method to deal with the
replicated gene symbols. It can be \code{"IQR"}, \code{"largest"}, or
\code{"average"}. The default is \code{"IQR"}. See details.}
}
\value{
A Study object with datasets of unique row entries
}
\description{
Title Handle Duplicated Gene Symbol
}

