% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.replicate.R
\name{PoolReplicate}
\alias{PoolReplicate}
\title{Pool the replicated gene symbols}
\usage{
PoolReplicate(dataset, method = REPLICATE.IQR)
}
\arguments{
\item{dataset}{A numeric matrix of gene expressions without
missing values or a Study object. For matrix, each row is for one gene
and each column is for one sample. The row names are gene symbols.
If dataset is Study object, the same rule applies to the matrix in
Study@datasets.}

\item{method}{A character string of the method to deal with the
replicated gene symbols. It can be \code{"IQR"}, \code{"largest"}, or
\code{"average"}. The default is \code{"IQR"}. See details.}
}
\value{
A numeric matrix of gene expressions without replicated
gene symbols. Each row is for one gene and each column is for one
sample. The row names are gene symbols.
}
\description{
Pool the replicated gene symbols to one using the method \code{"IQR"},
\code{"largest"}, or \code{"average"}.
}
\details{
\code{method} can be \code{"IQR"}, \code{"largest"}, or \code{"average"}.
\tabular{llll}{
\code{IQR} \tab \tab \tab \cr
\tab \tab \tab Represent the target gene symbol using the genes that
presented the greatest inter-quartile range (IQR). This method has been
recommended in Bioconductor.\cr
\code{largest} \tab \tab \tab \cr
\tab \tab \tab Represent the target gene symbol using the largest
expression.\cr
\code{average} \tab \tab \tab \cr
\tab \tab \tab Represent the target gene symbol using the average of
the expressions.
}
}
\examples{
data(study.eg)
data(preproc.option)
mydata <- Annotate(dataset=study.eg[[1]], id.type=ID.TYPE.probeID,
                   platform=PLATFORM.hgu133plus2)
mydata <- Impute(dataset=mydata)
res  <- PoolReplicate(dataset=mydata)
}
\author{
Lin Wang, Schwannden Kuo
}

