% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{print.Study}
\alias{print.Study}
\title{Title Custom print Function}
\usage{
\method{print}{Study}(x, ...)
}
\arguments{
\item{x}{a Study object.}

\item{...}{Any extra arguments}
}
\value{
show the meta data of the Study object.
}
\description{
Title Custom print Function
}
\examples{
data(preproc.option)
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=list(matrix()))
print(study)
}
\author{
Schwannden Kuo
}

