% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.expr.R
\name{ReadClinical}
\alias{ReadClinical}
\title{Title Read Clinical File}
\usage{
ReadClinical(file.name, sep = ",", quote = "\\"")
}
\arguments{
\item{file.name}{path to clinical file}

\item{sep}{the field separator character. Values on each line of
the file are separated by this character. If \code{sep = ""}, the
separator is 'white space', that is one or more spaces, tabs,
newlines or carriage returns. The default is \code{,}. See
\link{read.csv}.}

\item{quote}{the set of quoting characters. To disable quoting
altogether, use \code{quote = ""}.  See \link{scan} for the
behaviour on quotes embedded in quotes. Quoting is only
considered for columns read as character, which is all of them
unless colClasses is specified.  The default is \code{\"}. See
\link{read.csv}.}
}
\value{
a data.frame of clinical data
}
\description{
Title Read Clinical File
}
\examples{
data(study.eg)
data(preproc.option)
write.csv(study.eg@datasets[[1]], file="expr.csv")
write.csv(study.eg@clinicals[[1]], file="clinical.csv")
study <- ReadExpr(file.name="expr.csv", name="test", dtype=DTYPE.microarray)
clinical <- ReadClinical(file.name="clinical.csv")
study <- setClinical(study, list(clinical))
}

