% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.expr.R
\name{ReadExpr}
\alias{ReadExpr}
\title{Import and reformat gene expression data set}
\usage{
ReadExpr(file.name, name, dtype, clinical.file = NULL, log = TRUE,
  sep = ",", quote = "\\"", header = T)
}
\arguments{
\item{file.name}{a character string of the name of the file to be
read. See details.}

\item{name}{The name of the Study object returned}

\item{dtype}{The data type of the dataset (should be one of DTYPE.all)}

\item{clinical.file}{if there is a clinical file containing clinical 
data relevent to the dataset.}

\item{log}{logical. Specify whether the gene expressions need to
be log2-transformed. The default is \code{TRUE}.}

\item{sep}{the field separator character. Values on each line of
the file are separated by this character. If \code{sep = ""}, the
separator is 'white space', that is one or more spaces, tabs,
newlines or carriage returns. The default is \code{"\t"}. See
\link{read.csv}.}

\item{quote}{the set of quoting characters. To disable quoting
altogether, use \code{quote = ""}.  See \link{scan} for the
behaviour on quotes embedded in quotes. Quoting is only
considered for columns read as character, which is all of them
unless colClasses is specified.  The default is \code{'"'}. See
\link{read.csv}.}

\item{header}{a logical value indicating whether the file contains
the names of the variables as its first line. If missing, the
value is determined from the file format: \code{header} is set to
\code{TRUE} if and only if the first row contains one fewer field
than the number of columns. The default is \code{TRUE}.}
}
\value{
A Study object
}
\description{
Import gene expression data set into R and reformat it to the form
required for other functions in this package.
}
\details{
The file to be read should be prepared strictly according
following format. The first column should be gene ID and the
remaining columns should be the gene expressions for samples. The
first row should be sample names and the remaining rows
should be gene expression profiles for the samples.
}
\examples{
data(study.eg)
data(preproc.option)
write.csv(study.eg[[1]], file="study.txt")
mydata <- ReadExpr(file.name="study.txt", name="test", dtype=DTYPE.microarray)
}
\author{
Lin Wang, Schwannden Kuo
}

