% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_zero_variance.R
\name{RemoveZeroVariance}
\alias{RemoveZeroVariance}
\title{Remove zero variance columns from a dataframe.}
\usage{
RemoveZeroVariance(data, verbose = TRUE)
}
\arguments{
\item{data}{A dataframe or matrix}

\item{verbose}{If TRUE, prints names of removed variables}
}
\value{
A dataframe with the zero variance columns removed
}
\description{
Remove zero variance columns from a dataframe.
}
