% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{REPLICATE.all}
\alias{REPLICATE.all}
\title{Constant Option}
\format{An object of class \code{character} of length 3.}
\usage{
REPLICATE.all
}
\description{
Constant Option
}
\keyword{datasets}

