% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{REPLICATE.average}
\alias{REPLICATE.average}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
REPLICATE.average
}
\description{
Constant Option
}
\keyword{datasets}

