% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{setClinical}
\alias{setClinical}
\title{Title Set Clinical Data}
\usage{
setClinical(study, clinicals)
}
\arguments{
\item{study}{a Study object}

\item{clinicals}{a list of clinical data.frame. The length of clinicals, if none
empty, must equal the length of the datasets. And clinical[[i]] is the clinical
data for datasets[[i]]. Each clinical data.frame has row names corresponding 
to sample ID, and column names corresponding to clinical recording. And all 
samples in the corresponding dataset should have a clinical 
while clinical data can contain more then samples' clinical data.}
}
\value{
a study with updated clinical data
}
\description{
Title Set Clinical Data
}
\examples{
data(preproc.option)
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=list(matrix()))
setClinical(study, list())
}
\author{
Schwannden Kuo
}

