% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\docType{methods}
\name{show,Study-method}
\alias{show,Study-method}
\title{Title Custom show Function}
\usage{
\S4method{show}{Study}(object)
}
\arguments{
\item{object}{Study. A Study object.}
}
\value{
show the meta data of the Study object.
}
\description{
Title Custom show Function
}
\examples{
data(preproc.option)
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=list(matrix()))
show(study)
}
\author{
Schwannden Kuo
}

