% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{SPECIES.mouse}
\alias{SPECIES.mouse}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
SPECIES.mouse
}
\description{
Constant Option
}
\keyword{datasets}

