% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\docType{class}
\name{Study-class}
\alias{Study-class}
\title{Title Study Class}
\value{
A Study class
}
\description{
Title Study Class
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of the Study.}

\item{\code{dtype}}{character. Data type of the Study, should be one of DTYPE.all.}

\item{\code{ntype}}{character. Numeric type of the Study, should be one of NTYPE.all.}

\item{\code{stype}}{character. Study type, should be one of STYPE.all.}

\item{\code{datasets}}{list. A list of gene expression matrix.}

\item{\code{clinicals}}{list. A list of data.fram that is the clinical datas of the datasets.}
}}
\examples{
data(preproc.option)
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=list(matrix()))
}
\author{
Schwannden Kuo
}

