\name{study.eg}
\docType{data}
\alias{study.eg}
\title{An Study object with clinical data}
\description{
This is an annotated Study object. It is a single study with 30 features and 19 samples.
each sample has clinical data that is accessible in \code{study.eg@clinicals[[1]]}.
}
\usage{data(study.eg)}
\keyword{datasets}
