% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\docType{methods}
\name{stype,list-method}
\alias{stype,Study-method}
\alias{stype,list-method}
\title{stype method}
\usage{
\S4method{stype}{list}(object)

\S4method{stype}{Study}(object)
}
\arguments{
\item{object}{a list of dataset, or a Study object.}
}
\description{
stype method

stype method
}

