% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{stype}
\alias{stype}
\title{Title Generic Function: stype}
\usage{
stype(object)
}
\arguments{
\item{object}{a list of dataset, or a Study object.}
}
\value{
a string for the study type of the object.
}
\description{
Title Generic Function: stype
}
\examples{
data(preproc.option)
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=list(matrix()))
stype(study)
stype(list(matrix()))
}
\author{
Schwannden Kuo
}

