% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{STYPE.single}
\alias{STYPE.single}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
STYPE.single
}
\description{
Constant Option
}
\keyword{datasets}

