% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.option.R
\docType{data}
\name{TO.MATRIX.clinicals}
\alias{TO.MATRIX.clinicals}
\title{Constant Option}
\format{An object of class \code{character} of length 1.}
\usage{
TO.MATRIX.clinicals
}
\description{
Constant Option
}
\keyword{datasets}

