% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{to.matrix}
\alias{to.matrix}
\title{Title Study to matrix}
\usage{
to.matrix(study, what = TO.MATRIX.datasets)
}
\arguments{
\item{study}{A Study object.}

\item{what}{get datasets matrix or clinical data matrix. Datasets matrix 
results from datasets being column binded together. clinical data matrix
results from clinical data being row binded together.}
}
\value{
The Gene expression matrix of the Study, and if more then one dataset
present in the Study, the datasets are column binded.
}
\description{
Title Study to matrix
}
\examples{
data(study.eg)
data(preproc.option)
dataset.matrix <- to.matrix(study.eg)
clinical.matrix <- to.matrix(study.eg, what=TO.MATRIX.clinicals)
}
\author{
Schwannden Kuo
}

