\name{study.eg}
\docType{data}
\alias{study.eg}
\title{A toy data set for preprocessing}
\description{
This data set is a list containing three components \code{study1}, \code{study2},
and \code{study3} for three studies. Each components is a numeric matrix of gene
expressions. Each row is for one probe and each column is for one sample.
}
\usage{data(study.eg)}
\keyword{datasets}
