% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{to.matrix}
\alias{to.matrix}
\title{Title Study to matrix}
\usage{
to.matrix(study)
}
\arguments{
\item{study}{A Study object.}
}
\value{
The Gene expression matrix of the Study, and if more then one dataset
present in the Study, the datasets are column binded.
}
\description{
Title Study to matrix
}
\examples{
data(study.eg)
data(preproc.option)
study <- new("Study", name="test", dtype=DTYPE.microarray, datasets=study.eg)
to.matrix(study)
}
\author{
Schwannden Kuo
}

